+++ src/main/java/de/jeff_media/BestTools/PlayerSetting.java
@@ -46,6 +46,7 @@ public class PlayerSetting {
         }
 
         PlayerSetting(Player player, File file) {
+                //System.out.println("New PlayerSetting player, file");
                 this.player = player;
                 YamlConfiguration yaml = YamlConfiguration.loadConfiguration(file);
                 blacklist = new Blacklist(yaml.getStringList("blacklist"));
@@ -57,6 +58,24 @@ public class PlayerSetting {
                 this.swordOnMobs = yaml.getBoolean("swordOnMobs",true);
                 this.favoriteSlot = yaml.getInt("favoriteSlot",main.getConfig().getInt("favorite-slot"));
                 main.debug("Loaded player setting from file "+file.getPath());
+
+                getPDCValues(player);
+        }
+
+        private void getPDCValues(Player player) {
+                //System.out.println("getPDCValues player");
+                if(player.getPersistentDataContainer().has(DATA, DataType.FILE_CONFIGURATION)) {
+                        //System.out.println("Player has saved data, loading...");
+                        FileConfiguration conf = player.getPersistentDataContainer().get(DATA, DataType.FILE_CONFIGURATION);
+                        this.bestToolsEnabled = conf.getBoolean("bestToolsEnabled");
+                        //System.out.println("Enabled: " + bestToolsEnabled);
+                        this.hasSeenBestToolsMessage = conf.getBoolean("hasSeenBestToolsMessage");
+                        this.hasSeenRefillMessage = conf.getBoolean("hasSeenRefillMessage");
+                        this.refillEnabled = conf.getBoolean("refillEnabled");
+                        this.hotbarOnly = conf.getBoolean("hotbarOnly");
+                        //this.swordOnMobs = conf.getBoolean("swordOnMobs");
+                        //this.favoriteSlot = conf.getInt("favoriteSlot");
+                }
         }
 
         private static <T,Z> Z getPdc(Player player, NamespacedKey key, PersistentDataType<T,Z> type, Z defaultValue) {
@@ -64,6 +83,7 @@ public class PlayerSetting {
         }
 
         private void save() {
+                //System.out.println("Saving to PDC...");
                 FileConfiguration conf = new YamlConfiguration();
                 conf.set("blacklist",blacklist.toStringList());
                 conf.set("bestToolsEnabled",bestToolsEnabled);
@@ -77,6 +97,7 @@ public class PlayerSetting {
         }
 
         PlayerSetting(Player player, boolean bestToolsEnabled, boolean refillEnabled, boolean hotbarOnly, int favoriteSlot, boolean swordOnMobs) {
+
                 this.player = player;
                 this.blacklist = new Blacklist();
                 this.bestToolsEnabled = bestToolsEnabled;
@@ -85,9 +106,9 @@ public class PlayerSetting {
                 this.hasSeenRefillMessage = false;
                 this.hotbarOnly = hotbarOnly;
                 this.swordOnMobs= swordOnMobs;
-                this.changed = true;
                 this.favoriteSlot = favoriteSlot;
-
+                getPDCValues(player);
+                this.save();
         }
 
         Blacklist getBlacklist() {
@@ -96,57 +117,37 @@ public class PlayerSetting {
 
         boolean toggleBestToolsEnabled() {
                 bestToolsEnabled =!bestToolsEnabled;
-                changed = true;
+                save();
                 return bestToolsEnabled;
         }
 
         boolean toggleRefillEnabled() {
                 refillEnabled=!refillEnabled;
-                changed = true;
+                save();
                 return refillEnabled;
         }
 
         boolean toggleHotbarOnly() {
                 hotbarOnly=!hotbarOnly;
-                changed = true;
+                save();
                 return hotbarOnly;
         }
 
         void setHasSeenBestToolsMessage(boolean seen) {
                 if(seen== hasSeenBestToolsMessage) return;
                 hasSeenBestToolsMessage = seen;
-                changed = true;
+                save();
         }
 
         void setHasSeenRefillMessage(boolean seen) {
                 if(seen== hasSeenRefillMessage) return;
                 hasSeenRefillMessage = seen;
-                changed = true;
+                save();
         }
 
         void setFavoriteSlot(int favoriteSlot) {
                 this.favoriteSlot = favoriteSlot;
-                changed = true;
-        }
-
-        void save(Player player) {
-                //main.debug("Saving player setting to file "+file.getPath());
-                YamlConfiguration yaml = new YamlConfiguration();
-                yaml.set("bestToolsEnabled", bestToolsEnabled);
-                yaml.set("hotbarOnly",hotbarOnly);
-                yaml.set("refillEnabled",refillEnabled);
-                yaml.set("hasSeenBestToolsMessage", hasSeenBestToolsMessage);
-                yaml.set("hasSeenRefillMessage", hasSeenRefillMessage);
-                yaml.set("favoriteSlot",favoriteSlot);
-                yaml.set("swordOnMobs",swordOnMobs);
-                yaml.set("blacklist",blacklist.toStringList());
-                player.getPersistentDataContainer().set(DATA, DataType.FILE_CONFIGURATION,yaml);
-                /*try {
-                        yaml.save(file);
-                } catch (IOException e) {
-                        main.getLogger().warning("Error while saving playerdata file "+file.getName());
-                        //e.printStackTrace();
-                }*/
+                save();
         }
 
 }